refactor(app): ChooseRobotToRunProtocolSlideout->Desktop
