+++ app/src/organisms/Desktop/Devices/RobotOverflowMenu.tsx
@@ -26,7 +26,7 @@ import { CONNECTABLE, removeRobot } from '/app/redux/discovery'
 import { useIsRobotOnWrongVersionOfSoftware } from '/app/redux/robot-update'
 import { Divider } from '/app/atoms/structure'
 import { getTopPortalEl } from '/app/App/portal'
-import { ChooseProtocolSlideout } from '/app/organisms/ChooseProtocolSlideout'
+import { ChooseProtocolSlideout } from '/app/organisms/Desktop/ChooseProtocolSlideout'
 import { useCurrentRunId } from '/app/resources/runs'
 import { ConnectionTroubleshootingModal } from './ConnectionTroubleshootingModal'
 import { useIsRobotBusy } from './hooks'
